move shipping method to genaral query, add isActive support, code clean up
