+++ server/graphql/Cargo.toml
@@ -48,7 +48,6 @@ graphql_purchase_order_line = { path = "purchase_order_line" }
 graphql_goods_received = { path = "goods_received" }
 graphql_goods_received_line = { path = "goods_received_line" }
 graphql_contact = { path = "contact" }
-graphql_shipping_method = { path = "shipping_method" }
 
 actix-web = { workspace = true }
 async-graphql = { workspace = true }
