+++ server/graphql/general/src/queries/mod.rs
@@ -75,6 +75,7 @@ pub mod abbreviation;
 pub mod diagnosis;
 pub mod insurance_policies;
 pub mod insurance_providers;
+pub mod shipping_method;
 
 #[cfg(test)]
 mod tests;
